Try a different approach to always call useOnBlockProps
