+++ packages/block-library/src/paragraph/drop-zone.js
@@ -1,15 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { useSelect } from '@wordpress/data';
-import {
-	__experimentalUseOnBlockDrop as useOnBlockDrop,
-	store as blockEditorStore,
-} from '@wordpress/block-editor';
-import {
-	__experimentalUseDropZone as useDropZone,
-	useReducedMotion,
-} from '@wordpress/compose';
+import { useReducedMotion } from '@wordpress/compose';
 import { Popover, __unstableMotion as motion } from '@wordpress/components';
 
 const animateVariants = {
@@ -18,30 +10,7 @@ const animateVariants = {
 	exit: { opacity: 0, scaleY: 0.9 },
 };
 
-export default function DropZone( {
-	paragraphElement,
-	clientId,
-	setIsDropZoneVisible,
-} ) {
-	const { rootClientId, blockIndex } = useSelect(
-		( select ) => {
-			const selectors = select( blockEditorStore );
-			return {
-				rootClientId: selectors.getBlockRootClientId( clientId ),
-				blockIndex: selectors.getBlockIndex( clientId ),
-			};
-		},
-		[ clientId ]
-	);
-	const onBlockDrop = useOnBlockDrop( rootClientId, blockIndex, {
-		action: 'replace',
-	} );
-	const dropZoneRef = useDropZone( {
-		onDrop: onBlockDrop,
-		onDragLeave: () => {
-			setIsDropZoneVisible( false );
-		},
-	} );
+export default function DropZone( { paragraphElement } ) {
 	const reducedMotion = useReducedMotion();
 
 	return (
@@ -55,7 +24,6 @@ export default function DropZone( {
 			className="wp-block-paragraph__drop-zone"
 		>
 			<motion.div
-				ref={ dropZoneRef }
 				style={ {
 					width: paragraphElement?.offsetWidth,
 					height: paragraphElement?.offsetHeight,
